[MODFIN-166] - Move models to appropriate package
